docs: update tutorial
- add News::new() to show HTML form
- change News::view() to show()
- change routing
