+++ user_guide_src/source/tutorial/create_news_items/002.php
@@ -3,46 +3,18 @@
 namespace App\Controllers;
 
 use App\Models\NewsModel;
+use CodeIgniter\Exceptions\PageNotFoundException;
 
 class News extends BaseController
 {
     // ...
 
-    public function create()
+    public function new()
     {
         helper('form');
 
-        // Checks whether the form is submitted.
-        if (! $this->request->is('post')) {
-            // The form is not submitted, so returns the form.
-            return view('templates/header', ['title' => 'Create a news item'])
-                . view('news/create')
-                . view('templates/footer');
-        }
-
-        $post = $this->request->getPost(['title', 'body']);
-
-        // Checks whether the submitted data passed the validation rules.
-        if (! $this->validateData($post, [
-            'title' => 'required|max_length[255]|min_length[3]',
-            'body'  => 'required|max_length[5000]|min_length[10]',
-        ])) {
-            // The validation fails, so returns the form.
-            return view('templates/header', ['title' => 'Create a news item'])
-                . view('news/create')
-                . view('templates/footer');
-        }
-
-        $model = model(NewsModel::class);
-
-        $model->save([
-            'title' => $post['title'],
-            'slug'  => url_title($post['title'], '-', true),
-            'body'  => $post['body'],
-        ]);
-
         return view('templates/header', ['title' => 'Create a news item'])
-            . view('news/success')
+            . view('news/create')
             . view('templates/footer');
     }
 }
