+++ user_guide_src/source/tutorial/news_section/003.php
@@ -13,7 +13,7 @@ class News extends BaseController
         $data['news'] = $model->getNews();
     }
 
-    public function view($slug = null)
+    public function show($slug = null)
     {
         $model = model(NewsModel::class);
