+++ user_guide_src/source/tutorial/news_section/008.php
@@ -2,12 +2,11 @@
 
 // ...
 
-use App\Controllers\News;
+use App\Controllers\News; // Add this line
 use App\Controllers\Pages;
 
-$routes->get('news/(:segment)', [News::class, 'view']);
-$routes->get('news', [News::class, 'index']);
+$routes->get('news', [News::class, 'index']);           // Add this line
+$routes->get('news/(:segment)', [News::class, 'show']); // Add this line
+
 $routes->get('pages', [Pages::class, 'index']);
 $routes->get('(:segment)', [Pages::class, 'view']);
-
-// ...
