+++ user_guide_src/source/tutorial/static_pages/003.php
@@ -1,7 +1,8 @@
 <?php
 
-// ...
+use CodeIgniter\Router\RouteCollection;
 
+/**
+ * @var RouteCollection $routes
+ */
 $routes->get('/', 'Home::index');
-
-// ...
